wix-style-react | fix(AudioPlayer): audio file can be played twice in an edge case. (#5533)
Reason: "wasEverPlayed" is set to true too early (when someone hit play but before audio was really played), then it triggered another "useEffect" which requests for another play. Because audio is not yet played, the getter from howler (playing) returns false which triggers a second play.

Solution: set wasEverPlayed to true only when onPlay is called.